% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs_count.R
\name{jobs_count}
\alias{jobs_count}
\title{jobs_count Function}
\usage{
jobs_count(search.keyword, search.location)
}
\arguments{
\item{search.keyword}{Key word to search the site indeed.com.}

\item{search.location}{location of job search.}
}
\description{
This function returns the total number of Jobs from Indeed.com.
}
\examples{
jobs_count()
}
