% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs_df.R
\name{jobs_df}
\alias{jobs_df}
\title{jobs_df Function}
\usage{
jobs_df(search.keyword, search.location)
}
\arguments{
\item{search.keyword}{Key word to search the site indeed.com.}

\item{search.location}{location of job search.}
}
\description{
This function returns the list of Jobs from Indeed.com in a dataframe.
}
